Move new types files out of their folders to their modules folders
