+++ src/email/types.ts
@@ -1,5 +1,5 @@
 import type nodemailer from "nodemailer";
-import type { WithRequiredProperties } from "../../types/types";
+import type { WithRequiredProperties } from "../types";
 
 export type EmailOptions = WithRequiredProperties<
   nodemailer.SendMailOptions,
